+++ src/tasks/bamboo-test.js
@@ -1,13 +1,18 @@
 /* eslint no-var: 0 */
 
-var gulp = require('gulp');
-var mocha = require('gulp-spawn-mocha');
+var util = require('gulp-util');
+var test = require('./test');
 
+/**
+ * @deprecated
+ */
 module.exports = function(config) {
-    return gulp
-        .src(config.testEntryPoint)
-        .pipe(mocha({
-            R: 'mocha-bamboo-reporter',
-            compilers: 'jsx?:babel-core/register',
-        }));
+
+    util.log(
+        util.colors.yellow('[DEPRECATION WARNING]'),
+        'You used the gulp `bamboo-test` task. Please use `test` instead.'
+    );
+
+    return test(config);
+
 };