+++ src/util/HTMLTemplatePlugin.js
@@ -4,6 +4,10 @@ var _ = require('lodash');
 
 var cssTemplate = _.template('    <link rel="stylesheet" type="text/css" href="${styleSheet}">\n');
 
+/**
+ * @deprecated
+ * @constructor
+ */
 function HTMLTemplatePlugin() {}
 
 module.exports = HTMLTemplatePlugin;